feat: moved constants in their own file and changed the army background to be animated
